+++ src/components/popover/popover.js
@@ -153,7 +153,9 @@ export class EuiPopover extends Component {
       this.setState({ suppressingPopover: false, isOpening: true }); // eslint-disable-line react/no-did-mount-set-state
     }
 
-    window.addEventListener('scroll', this.positionPopover);
+    if (this.props.repositionOnScroll) {
+      window.addEventListener('scroll', this.positionPopover);
+    }
 
     this.updateFocus();
   }
@@ -169,6 +171,15 @@ export class EuiPopover extends Component {
           isOpening: true,
         });
       });
+    }
+
+    // update scroll listener
+    if (prevProps.repositionOnScroll !== this.props.repositionOnScroll) {
+      if (this.props.repositionOnScroll) {
+        window.addEventListener('scroll', this.positionPopover);
+      } else {
+        window.removeEventListener('scroll', this.positionPopover);
+      }
     }
 
     // The popover is being closed.
@@ -424,6 +435,7 @@ EuiPopover.propTypes = {
     PropTypes.node,
     PropTypes.instanceOf(HTMLElement)
   ]),
+  repositionOnScroll: PropTypes.bool,
 };
 
 EuiPopover.defaultProps = {